requirements.txt
anyio==4.4.0
APScheduler==3.6.3
Brotli==1.1.0
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mutagen==1.47.0
pycryptodomex==3.20.0
python-telegram-bot==13.7
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
tzlocal==5.2
urllib3==2.2.2
websockets==12.0
yt-dlp==2024.7.25